lmb-1214: created generated-form model that inherits from form model
